+++ lib/cuckoo/core/database.py
@@ -1314,7 +1314,9 @@ class Database(object, metaclass=Singleton):
         Handles ZIP file submissions, submitting each extracted file to the database
         Returns a list of added task IDs
         """
+        task_id = False
         task_ids = []
+        config = False
         sample_parent_id = None
         # force auto package for linux files
         if platform == "linux":
@@ -1338,7 +1340,6 @@ class Database(object, metaclass=Singleton):
 
         # create tasks for each file in the archive
         for file in extracted_files:
-            config = False
             if static:
                 config = static_extraction(file)
                 if config:
@@ -1366,10 +1367,11 @@ class Database(object, metaclass=Singleton):
                     tlp=tlp,
                     source_url=source_url,
                 )
-                if task_id:
-                    task_ids.append(task_id)
+            if task_id:
+                task_ids.append(task_id)
 
-        return task_ids
+        # this is aim to return custom data, think of this as kwargs
+        return task_ids, {"config": config}
 
     @classlock
     def add_pcap(